Refactor: Separate out task services
